+++ backend/app/modules/auth/enums.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class ProviderEnum(str, Enum):
+    google = "google"
+    kakao = "kakao"
+    naver = "naver"
+
+
+class UserRoleEnum(str, Enum):
+    admin = "admin"
+    user = "user"
